feat: refactor event management feature structure
