+++ src/features/events/index.ts
@@ -1,0 +1,8 @@
+export { default as EventManagementPage } from '@/features/events/views/EventManagementPage';
+export * from '@/features/events/types/event';
+export * from '@/features/events/hooks/useEventCreationForm';
+export * from '@/features/events/hooks/useEventFilters';
+export * from '@/features/events/hooks/useMyEvents';
+export * from '@/features/events/components/EventCreationForm';
+export * from '@/features/events/components/EventList';
+export * from '@/features/events/components/MyEventsList';
